feat(phase-2): float-in-regulated-code lint rule + fix a real instance it found
Phase 2 backbone ratchet item 5, optional (PHASE_2_BACKBONE.md Sec 4). ADR-0007 / Doc 110 CALC-FR-001 /
CLAUDE.md Sec 9: "No binary float for a regulated quantity."

Full generality (flag every `float` field) isn't safe to automate: engineering timing/duration floats
(retry_after_seconds, timeout_seconds, backoff_seconds, ...) are pervasive and legitimate throughout
app/modules/erp/reliability.py and similar files, and can't be told apart from a regulated quantity by
field name alone. Added a 5th, deliberately narrower check to the guardrails tooling built for item 3/6
instead -- `no-float-for-decimal-column` -- with zero false-positive risk: it flags only an ORM column
whose database type is explicitly Numeric/DECIMAL (the author's own declared intent: exact decimal
storage) but whose Mapped[...] annotation says float instead of Decimal, a self-contradictory shape.

That check found one real bug: qms.NcrDisposition.quantity was Mapped[float] over a Numeric(18, 6)
column -- SQLAlchemy handed the application a lossy binary float on every read despite the column being
stored as an exact decimal, silently defeating DATA-FR-019 downstream of the DB layer. Its
DispositionNcrCommand.quantity Pydantic counterpart had the same float typing. Both fixed to Decimal,
matching the plain-`Decimal`-field convention already used everywhere else in this codebase
(CTR-FR-015 decimal-as-string transport) -- no arithmetic or serialization code elsewhere touches this
field beyond direct pass-through, confirmed before changing it.

Also added 2 negative-fixture tests (violation + adjacent allowed pattern, same style as the other 4
checks) and documented the new check in the README.

Verified: tooling/guardrails/validate.py clean against app/ (5/5 checks, 0 findings -- the ncr_models.py
fix was the only instance); 11/11 guardrails pytest tests pass; tests/test_qms_ncr.py 16/16 passed
against the freshly rebuilt test database (item 4).

+++ services/gxp-api/app/modules/qms/ncr_commands.py
@@ -7,6 +7,7 @@ Documents 26/27 (see ncr_models.py's module docstring and docs/generated/18_SPEC
 
 import uuid
 from datetime import datetime, timezone
+from decimal import Decimal
 
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -276,7 +277,7 @@ class DispositionNcrCommand(CommandEnvelope):
     disposition_type: str
     affected_scope: list[dict]
     justification: str
-    quantity: float | None = None
+    quantity: Decimal | None = None
     serials: list[str] | None = None
     rework_route: dict | None = None
     follow_up_test_requirements: dict | None = None
